+++ app/layout.tsx
@@ -1,3 +1,6 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import { ToastContainer } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 import ReactQueryProvider from '@/hooks/queries/ReactQueryProvider';
 import 'styles/global.css';
 
@@ -9,6 +12,7 @@ export default function RootLayout({
   return (
     <html lang="ko">
       <ReactQueryProvider>
+        <ToastContainer />
         <body>{children}</body>
       </ReactQueryProvider>
     </html>
