+++ package.json
@@ -24,6 +24,7 @@
     "react-focus-lock": "^2.9.5",
     "react-hook-form": "^7.46.2",
     "react-intersection-observer": "^9.5.2",
+    "react-toastify": "^9.1.3",
     "socket.io-client": "^4.7.2",
     "swiper": "9.4.1",
     "tailwind-scrollbar-hide": "^1.1.7",
